requirements.txt
# Web framework
flask==3.0.3
flask-migrate==4.0.7

# Database
sqlalchemy==2.0.36
flask-sqlalchemy==3.1.1
pyodbc==5.1.0
alembic==1.13.3

# Azure AI
azure-ai-projects==1.0.0b11
azure-identity==1.19.0

# HTTP
requests==2.32.3

# Environment
python-dotenv==1.0.1

# File reading
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5